fix: update data paths in configuration and scripts for consistency
Please review this commit carefully. In opinion, from the given README.md, the data should be stored in `vnegnn/data/data/` for experiment=VNEGNN whereas the data for experiment=equipocket should be stored in `vnegnn/data/equipocket`.
This commit adapts the paths in the preprocessing scripts as well as the hydro config accordingly.